+++ app/components/omnibox/directives/search-directive.js
@@ -35,21 +35,16 @@ angular.module('omnibox')
      * with the right query and puts in on the scope.
      */
     scope.search = function () {
-      if (scope.geoquery && scope.geoquery.length > 1) {
-        LocationService.search(scope.geoquery)
-          .then(function (response) {
-            scope.geoquery = "";
-            if (response.length !== 0) {
-              scope.box.content.location = {
-                data: response
-              };
-            }
+      LocationService.search(scope.geoquery, State.spatial)
+        .then(function (response) {
+          if (response.status === LocationService.ggStatus.ok) {
+            scope.box.content.location = response.results;
           }
-        );
-      }
-      else {
-        scope.geoquery = "";
-      }
+          else {
+            scope.geoquery = "";
+          }
+        }
+      );
     };
 
     /**
@@ -73,7 +68,6 @@ angular.module('omnibox')
      * (5) - Clear the click feedback.
      */
     scope.cleanInput = function () {
-
       State.box.type = "point";
       scope.geoquery = "";
       scope.box.content = {};
@@ -86,19 +80,10 @@ angular.module('omnibox')
      * @description zooms to search result
      * @param {object} one search result.
      */
-    scope.zoomTo = function (obj) {
-      if (obj.boundingbox) {
-        var southWest = new L.LatLng(obj.boundingbox[0], obj.boundingbox[2]);
-        var northEast = new L.LatLng(obj.boundingbox[1], obj.boundingbox[3]);
-        var bounds = new L.LatLngBounds(southWest, northEast);
-        MapService.fitBounds(bounds);
-      } else {
-        if (window.JS_DEBUG) {
-          throw new Error('Oops, no boundingbox on this result - TODO: show a proper message instead of this console error...');
-        }
-      }
+    scope.zoomTo = function (location) {
       destroyLocationModel();
       scope.cleanInput();
+      LocationService.zoomToResult(location);
     };
 
   };